% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_things.R
\name{add_two}
\alias{add_two}
\title{Function to add two things.}
\usage{
add_two(x, y)
}
\arguments{
\item{x}{Numeric value.}

\item{y}{Numeric value.}
}
\value{
A numeric value representing the sum of the values provided.
}
\description{
Function to add two things.
}
\examples{
add_two(1, 2)
}
