% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniquify_things.R
\name{clean}
\alias{clean}
\title{Find the unique values in a vector}
\usage{
clean(vec)
}
\arguments{
\item{vec}{Input character vector.}
}
\value{
A character vector.
}
\description{
Find the unique values in a vector
}
\examples{
fruit <- c("apple", "pear", "banana", "apple")
uniquify(fruit)
}
